fix some issues for regresson tests for python3

+++ regress/test_opcodes/test_RET.py
@@ -88,8 +88,8 @@ class test_RET_old_%06x_new_%06x(base_RET):
 #
 code = ''
 
-for old_pc in (0,255,256,(8*1024/2-1)):
-	for new_pc in (0,1,2,3,255,256,(8*1024/2-1)):
+for old_pc in (0,255,256,int(8*1024/2-1)):
+	for new_pc in (0,1,2,3,255,256,int(8*1024/2-1)):
 		args = (old_pc,new_pc)*4
 		code += template % args
 exec(code)
