+++ regress/test_opcodes/test_RETI.py
@@ -101,8 +101,8 @@ class test_RETI_old_%06x_new_%06x(base_RETI):
 #
 code = ''
 
-for old_pc in (0,255,256,(8*1024/2-1)):
-	for new_pc in (0,1,2,3,255,256,(8*1024/2-1)):
+for old_pc in (0,255,256,int(8*1024/2-1)):
+	for new_pc in (0,1,2,3,255,256,int(8*1024/2-1)):
 		args = (old_pc,new_pc)*4
 		code += template % args
 exec(code)
